# openapi.yaml
openapi: "3.0.0"

info:
  title: "Simple TextBlob API"
  description: "API that accepts a text string and uses the TextBlob python library to analyze it. Various endpoints return part-of-speech taggs, sentiment scores, spellchecked words, and translations of the input into other languages."
  version: "1.0"

paths:
  /ping:
    get:
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string


  /TagPOS:
    post:
      description: "Return the part-of-speech tags for the input text"

      requestBody:
        description: "Json with single field containing text to tag"
        required: true
        content:
          application/json: {}

      responses:
        '200':
          description: "Returns list of each word and its tag"
          content: 
           application/json: {}


  /CalcSentimentScore:
    post:
      description: "Calculates the sentiment polarity score for the input text"

      requestBody:
        description: "Json with single field containing text to analyze for sentiment"
        required: true
        content:
          application/json: {}

      responses:
        '200':
          description: "Returns a single number representing sentiment polarity"
          content: 
           application/json: {}


  /AutocorrectText:
    post:
      description: "Attempts to auto-correct any spelling mistakes in the text"

      requestBody:
        description: "Json with single field containing text to correct spelling"
        required: true
        content:
          application/json: {}

      responses:
        '200':
          description: "Returns best attempt at spelling-corrected text"
          content: 
           application/json: {}


  /TranslateText:
    post:
      description: "Translate input text into a different language (language codes: https://cloud.google.com/translate/docs/languages)"

      parameters:
        - in: "path"
          name: "toLang"
          schema:
            type: string
          required: true
          description: "Two-letter language code to translate to"

      requestBody:
        description: "Json with single field containing text to translate"
        required: true
        content:
          application/json: {}

      responses:
        '200':
          description: "Returns translated text"
          content: 
           application/json: {}